% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_checklists.R
\name{all_checklists}
\alias{all_checklists}
\title{All Checklists}
\usage{
all_checklists(board_name, save = FALSE)
}
\arguments{
\item{board_name}{The name of the board you want to browse.}

\item{save}{Set to TRUE to save a copy of the card information list.}
}
\description{
This function allows you to obtain checklists and card info for all cards on a chosen board.
}
\keyword{repello}
