% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards_info.R
\name{cards_info}
\alias{cards_info}
\title{Get Cards Information}
\usage{
cards_info(board_id, user_token = NULL)
}
\arguments{
\item{board_id}{The ID of the board you want to browse.}

\item{user_token}{The user token for an individual's Trello account.}
}
\description{
This function allows you to obtain the cards and the date of most recent modification for the cards on a specified board.
}
\keyword{repello}
