% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_new_tasks.R
\name{check_new_tasks}
\alias{check_new_tasks}
\title{Check New Tasks}
\usage{
check_new_tasks(list1, list2)
}
\arguments{
\item{list1}{A checklist for a Trello card.}

\item{list2}{A more recent checklist for the same card.}
}
\description{
This function allows you to compare checklists for new items and completions.
}
\keyword{repello}
