% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_checklists.R
\name{compare_checklists}
\alias{compare_checklists}
\title{Compare Checklists}
\usage{
compare_checklists(old_list, new_list)
}
\arguments{
\item{old_list}{A  prior list of Trello activities.}

\item{new_list}{A more recent list of Trello activities.}
}
\description{
This function compares checklists for all items in a list of lists.
}
\keyword{repello}
