% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_board_id.R
\name{get_board_id}
\alias{get_board_id}
\title{Get Board ID}
\usage{
get_board_id(board_name)
}
\arguments{
\item{board_name}{The name of the board you want to browse.}
}
\description{
This function allows you to obtain the ID of the Trello board you want to view.
}
\keyword{repello}
