% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_checklist.R
\name{get_checklist}
\alias{get_checklist}
\title{Get Checklist}
\usage{
get_checklist(board_id, card_name)
}
\arguments{
\item{board_id}{The ID of the board you want to browse.}

\item{card_name}{The name of the card you want to view}
}
\description{
This function allows you to obtain a checklist for a card of interest.
}
\keyword{repello}
