% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_key_token.R
\name{set_key_token}
\alias{set_key_token}
\title{Set Key-Token}
\usage{
set_key_token(key_file = NULL, token_file = NULL, path = getwd())
}
\arguments{
\item{key_file}{The name of the file which contains the Trello key}

\item{token_file}{The name of the file which contains the user token}

\item{path}{The path to the location of the user token file}
}
\description{
This function allows you to set the user key and token for API calls.  You
can either link to text files which contain the key and token, or you
can leave the arguments blank and manually input the key and token when prompted
}
\keyword{repello}
