% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_token.R
\name{set_token}
\alias{set_token}
\title{Set Token}
\usage{
set_token(filename, path = getwd())
}
\arguments{
\item{filename}{The name of the file which contains the user token.}

\item{path}{The path to the location of the user token file}
}
\description{
This function allows you to set the path to the user token file.
}
\keyword{repello}
