% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trello_updates.R
\name{trello_updates}
\alias{trello_updates}
\title{Trello Updates}
\usage{
trello_updates(
  board_name,
  prior = FALSE,
  recent = FALSE,
  save = FALSE,
  user_token = NULL
)
}
\arguments{
\item{board_name}{The name of the board you want to browse.}

\item{prior}{The name of the prior Trello list you want to compare to.  Defaults to most recently saved Trello object prior to current save.}

\item{recent}{The name of the Trello board to compare to a prior Trello object.  By default will use current object grab as most recent object.}

\item{save}{Set to TRUE to save a copy of the current Trello object.}

\item{user_token}{The user token for an individual's Trello account.}
}
\description{
This function allows you to check for updates to all cards on a Trello board, comparing to a prior board setting.
}
\keyword{repello}
